Refactor name to keyspace extension
