+++ README.test-harness.md
@@ -25,7 +25,7 @@ To create a local test database docker container, execute the following steps:
 First you need to build project - `mvn package` will do the job.
 
 ##### from IDE
-From your IDE, right click on the `liquibase.ext.cassandra.LiquibaseHarnessSuiteIT` test class present in `src/test/groovy` directory. 
+From your IDE, right click on the `liquibase.ext.keyspace.LiquibaseHarnessSuiteIT` test class present in `src/test/groovy` directory. 
 Doing so, will allow you to execute all the standard change object tests in the liquibase-test-harness as well as the
 Cassandra specific change objects tests created exclusively to test this extension (You can find this in the 
 `src/test/resources/liquibase/harness/change/changelogs/cassandra` directory).
