--- src/main/java/liquibase/ext/cassandra/datatype/CassandraTextDataType.java
@@ -1,30 +1,0 @@
-package liquibase.ext.cassandra.datatype;
-
-import liquibase.change.core.LoadDataChange;
-import liquibase.database.Database;
-import liquibase.datatype.DataTypeInfo;
-import liquibase.datatype.DatabaseDataType;
-import liquibase.datatype.LiquibaseDataType;
-import liquibase.ext.cassandra.database.CassandraDatabase;
-
-@DataTypeInfo(name = "text", minParameters = 0, maxParameters = 0, priority = LiquibaseDataType.PRIORITY_DATABASE)
-public class CassandraTextDataType extends LiquibaseDataType {
-
-    @Override
-    public boolean supports(Database database) {
-        return database instanceof CassandraDatabase;
-    }
-
-    @Override
-    public DatabaseDataType toDatabaseDataType(Database database) {
-        if (database instanceof CassandraDatabase) {
-            return new DatabaseDataType("TEXT", getParameters());
-        }
-        return super.toDatabaseDataType(database);
-    }
-
-    @Override
-    public LoadDataChange.LOAD_DATA_TYPE getLoadTypeName() {
-        return LoadDataChange.LOAD_DATA_TYPE.STRING;
-    }
-}
