--- src/main/java/liquibase/ext/cassandra/sqlgenerator/RenameColumnGeneratorCassandra.java
@@ -1,34 +1,0 @@
-package liquibase.ext.cassandra.sqlgenerator;
-
-import liquibase.database.Database;
-import liquibase.ext.cassandra.database.CassandraDatabase;
-import liquibase.sql.Sql;
-import liquibase.sql.UnparsedSql;
-import liquibase.sqlgenerator.SqlGeneratorChain;
-import liquibase.sqlgenerator.core.RenameColumnGenerator;
-import liquibase.statement.core.RenameColumnStatement;
-
-public class RenameColumnGeneratorCassandra extends RenameColumnGenerator {
-    @Override
-    public int getPriority() {
-        return PRIORITY_DATABASE;
-    }
-
-    @Override
-    public boolean supports(RenameColumnStatement statement, Database database) {
-        return database instanceof CassandraDatabase;
-    }
-
-    @Override
-    public Sql[] generateSql(RenameColumnStatement statement, Database database, SqlGeneratorChain sqlGeneratorChain) {
-        return new Sql[]{
-                new UnparsedSql(
-                        "ALTER TABLE "
-                                + database.escapeTableName(statement.getCatalogName(), statement.getSchemaName(), statement.getTableName())
-                                + " RENAME "
-                                + database.escapeColumnName(statement.getCatalogName(), statement.getSchemaName(), statement.getTableName(), statement.getOldColumnName())
-                                + " TO "
-                                + database.escapeColumnName(statement.getCatalogName(), statement.getSchemaName(), statement.getTableName(), statement.getNewColumnName()),
-                        getAffectedOldColumn(statement), getAffectedNewColumn(statement))};
-    }
-}
