+++ src/main/java/liquibase/ext/keyspace/sqlgenerator/CreateDatabaseChangeLogLockTableGeneratorCassandra.java
@@ -1,0 +1,35 @@
+package liquibase.ext.keyspace.sqlgenerator;
+
+import liquibase.database.Database;
+import liquibase.ext.keyspace.database.CassandraDatabase;
+import liquibase.sql.Sql;
+import liquibase.sqlgenerator.SqlGeneratorChain;
+import liquibase.sqlgenerator.SqlGeneratorFactory;
+import liquibase.sqlgenerator.core.CreateDatabaseChangeLogLockTableGenerator;
+import liquibase.statement.core.CreateDatabaseChangeLogLockTableStatement;
+import liquibase.statement.core.RawSqlStatement;
+
+public class CreateDatabaseChangeLogLockTableGeneratorCassandra extends CreateDatabaseChangeLogLockTableGenerator {
+
+    @Override
+    public int getPriority() {
+        return PRIORITY_DATABASE;
+    }
+
+    @Override
+    public boolean supports(CreateDatabaseChangeLogLockTableStatement statement, Database database) {
+        return super.supports(statement, database) && database instanceof CassandraDatabase;
+    }
+
+    @Override
+    public Sql[] generateSql(CreateDatabaseChangeLogLockTableStatement statement, Database database, SqlGeneratorChain sqlGeneratorChain) {
+
+        RawSqlStatement createTableStatement = new RawSqlStatement("CREATE TABLE IF NOT EXISTS " +
+                database.escapeTableName(database.getLiquibaseCatalogName(), database.getLiquibaseSchemaName(), "databasechangeloglock") +
+                " (ID INT, LOCKED BOOLEAN, LOCKGRANTED timestamp, LOCKEDBY TEXT, PRIMARY KEY (ID))");
+
+        return SqlGeneratorFactory.getInstance().generateSql(createTableStatement, database);
+
+    }
+
+}
