+++ src/main/java/liquibase/ext/keyspace/sqlgenerator/CreateTableGeneratorCassandra.java
@@ -1,0 +1,30 @@
+package liquibase.ext.keyspace.sqlgenerator;
+
+import liquibase.database.Database;
+import liquibase.ext.keyspace.database.CassandraDatabase;
+import liquibase.sql.Sql;
+import liquibase.sqlgenerator.SqlGeneratorChain;
+import liquibase.sqlgenerator.core.CreateTableGenerator;
+import liquibase.statement.core.CreateTableStatement;
+
+public class CreateTableGeneratorCassandra extends CreateTableGenerator {
+
+    @Override
+    public int getPriority() {
+        return PRIORITY_DATABASE;
+    }
+
+    @Override
+    public boolean supports(CreateTableStatement statement, Database database) {
+        return database instanceof CassandraDatabase;
+    }
+    @Override
+    public Sql[] generateSql(CreateTableStatement statement, Database database, SqlGeneratorChain sqlGeneratorChain) {
+        // Cassandra doesn't support not null constraints
+        statement.getNotNullColumns().clear();
+
+        return super.generateSql(statement, database, sqlGeneratorChain);
+
+    }
+
+}
