+++ src/main/java/liquibase/ext/keyspace/sqlgenerator/GetNextChangeSetSequenceValueGeneratorCassandra.java
@@ -1,0 +1,20 @@
+package liquibase.ext.keyspace.sqlgenerator;
+
+import liquibase.database.Database;
+import liquibase.ext.keyspace.database.CassandraDatabase;
+import liquibase.sqlgenerator.core.GetNextChangeSetSequenceValueGenerator;
+import liquibase.statement.core.GetNextChangeSetSequenceValueStatement;
+
+public class GetNextChangeSetSequenceValueGeneratorCassandra extends GetNextChangeSetSequenceValueGenerator {
+
+    @Override
+    public int getPriority() {
+        return PRIORITY_DATABASE;
+    }
+
+    @Override
+    public boolean supports(GetNextChangeSetSequenceValueStatement statement, Database database) {
+        return super.supports(statement, database) && database instanceof CassandraDatabase;
+    }
+
+}
