+++ src/main/java/liquibase/ext/keyspace/sqlgenerator/LockDatabaseChangeLogGeneratorCassandra.java
@@ -1,0 +1,32 @@
+package liquibase.ext.keyspace.sqlgenerator;
+
+import liquibase.database.Database;
+import liquibase.ext.keyspace.database.CassandraDatabase;
+import liquibase.sql.Sql;
+import liquibase.sqlgenerator.SqlGeneratorChain;
+import liquibase.sqlgenerator.SqlGeneratorFactory;
+import liquibase.sqlgenerator.core.LockDatabaseChangeLogGenerator;
+import liquibase.statement.core.LockDatabaseChangeLogStatement;
+import liquibase.statement.core.RawSqlStatement;
+
+public class LockDatabaseChangeLogGeneratorCassandra extends LockDatabaseChangeLogGenerator {
+
+    @Override
+    public int getPriority() {
+        return PRIORITY_DATABASE;
+    }
+
+    @Override
+    public boolean supports(LockDatabaseChangeLogStatement statement, Database database) {
+        return super.supports(statement, database) && database instanceof CassandraDatabase;
+    }
+
+    @Override
+    public Sql[] generateSql(LockDatabaseChangeLogStatement statement, Database database, SqlGeneratorChain sqlGeneratorChain) {
+        RawSqlStatement updateStatement = new RawSqlStatement("UPDATE " +
+                database.escapeTableName(database.getLiquibaseCatalogName(), database.getLiquibaseSchemaName(), "databasechangeloglock") +
+                " SET LOCKED = TRUE, LOCKEDBY = '" + hostname + " (" + hostaddress + ")" + "', LOCKGRANTED = " + System.currentTimeMillis() + " WHERE ID = 1");
+        return SqlGeneratorFactory.getInstance().generateSql(updateStatement, database);
+    }
+
+}
