+++ src/main/resources/META-INF/services/liquibase.changelog.ChangeLogHistoryService
@@ -1,1 +1,1 @@
-liquibase.ext.cassandra.changelog.CassandraChangeLogHistoryService
+liquibase.ext.keyspace.changelog.CassandraChangeLogHistoryService
