+++ src/main/resources/META-INF/services/liquibase.lockservice.LockService
@@ -1,1 +1,1 @@
-liquibase.ext.cassandra.lockservice.LockServiceCassandra
+liquibase.ext.keyspace.lockservice.LockServiceCassandra
