+++ src/main/resources/META-INF/services/liquibase.snapshot.SnapshotGenerator
@@ -1,4 +1,4 @@
-liquibase.ext.cassandra.snapshot.TableSnapshotGeneratorCassandra
-liquibase.ext.cassandra.snapshot.ColumnSnapshotGeneratorCassandra
-liquibase.ext.cassandra.snapshot.UniqueConstraintSnapshotGeneratorCassandra
-liquibase.ext.cassandra.snapshot.IndexSnapshotGeneratorCassandra
+liquibase.ext.keyspace.snapshot.TableSnapshotGeneratorCassandra
+liquibase.ext.keyspace.snapshot.ColumnSnapshotGeneratorCassandra
+liquibase.ext.keyspace.snapshot.UniqueConstraintSnapshotGeneratorCassandra
+liquibase.ext.keyspace.snapshot.IndexSnapshotGeneratorCassandra
