+++ src/main/resources/META-INF/services/liquibase.sqlgenerator.SqlGenerator
@@ -1,11 +1,11 @@
-liquibase.ext.cassandra.sqlgenerator.CreateDatabaseChangeLogLockTableGeneratorCassandra
-liquibase.ext.cassandra.sqlgenerator.CreateDatabaseChangeLogTableGeneratorCassandra
-liquibase.ext.cassandra.sqlgenerator.GetNextChangeSetSequenceValueGeneratorCassandra
-liquibase.ext.cassandra.sqlgenerator.LockDatabaseChangeLogGeneratorCassandra
-liquibase.ext.cassandra.sqlgenerator.UnlockDatabaseChangeLogGeneratorCassandra
-liquibase.ext.cassandra.sqlgenerator.DeleteGeneratorCassandra
-liquibase.ext.cassandra.sqlgenerator.InitializeDatabaseChangeLogLockTableGeneratorCassandra
-liquibase.ext.cassandra.sqlgenerator.CreateTableGeneratorCassandra
-liquibase.ext.cassandra.sqlgenerator.DropColumnGeneratorCassandra
-liquibase.ext.cassandra.sqlgenerator.RenameColumnGeneratorCassandra
-liquibase.ext.cassandra.sqlgenerator.TagDatabaseGeneratorCassandra
+liquibase.ext.keyspace.sqlgenerator.CreateDatabaseChangeLogLockTableGeneratorCassandra
+liquibase.ext.keyspace.sqlgenerator.CreateDatabaseChangeLogTableGeneratorCassandra
+liquibase.ext.keyspace.sqlgenerator.GetNextChangeSetSequenceValueGeneratorCassandra
+liquibase.ext.keyspace.sqlgenerator.LockDatabaseChangeLogGeneratorCassandra
+liquibase.ext.keyspace.sqlgenerator.UnlockDatabaseChangeLogGeneratorCassandra
+liquibase.ext.keyspace.sqlgenerator.DeleteGeneratorCassandra
+liquibase.ext.keyspace.sqlgenerator.InitializeDatabaseChangeLogLockTableGeneratorCassandra
+liquibase.ext.keyspace.sqlgenerator.CreateTableGeneratorCassandra
+liquibase.ext.keyspace.sqlgenerator.DropColumnGeneratorCassandra
+liquibase.ext.keyspace.sqlgenerator.RenameColumnGeneratorCassandra
+liquibase.ext.keyspace.sqlgenerator.TagDatabaseGeneratorCassandra
