--- src/test/groovy/liquibase/ext/cassandra/LiquibaseHarnessSuiteIT.groovy
@@ -1,6 +1,0 @@
-package liquibase.ext.cassandra
-
-import liquibase.harness.BaseHarnessSuite
-
-class LiquibaseHarnessSuiteIT extends BaseHarnessSuite {
-}
