--- src/test/groovy/liquibase/ext/cassandra/changelog/CassandraChangeLogHistoryServiceTest.groovy
@@ -1,6 +1,0 @@
-package liquibase.ext.cassandra.changelog
-
-import spock.lang.Specification
-
-class CassandraChangeLogHistoryServiceTest extends Specification {
-}
