+++ src/test/groovy/liquibase/ext/keyspace/LiquibaseHarnessSuiteIT.groovy
@@ -1,0 +1,6 @@
+package liquibase.ext.keyspace
+
+import liquibase.harness.BaseHarnessSuite
+
+class LiquibaseHarnessSuiteIT extends BaseHarnessSuite {
+}
