+++ src/test/groovy/liquibase/ext/keyspace/changelog/CassandraChangeLogHistoryServiceTest.groovy
@@ -1,0 +1,6 @@
+package liquibase.ext.keyspace.changelog
+
+import spock.lang.Specification
+
+class CassandraChangeLogHistoryServiceTest extends Specification {
+}
